+++ test/dummy/app/controllers/moveable_docs_controller.rb
@@ -13,12 +13,4 @@ class MoveableDocsController < ApplicationController
 
   def api
   end
-
-  def scalar
-    render layout: false
-  end
-
-  def openapi
-    render json: RecordingStudioApi.openapi_document(version: "v1")
-  end
 end